Reorganize Metafunctions
Add an api module
