+++ metafunctions/tests/test_call_state.py
@@ -1,7 +1,7 @@
 
 from metafunctions.tests.util import BaseTestCase
-from metafunctions.util import bind_call_state
-from metafunctions.util import node
+from metafunctions.api import bind_call_state
+from metafunctions.api import node
 from metafunctions.core import CallState
